Split flatsharp into codegen and runtime parts.
