+++ src/FlatSharp.Runtime/Attributes/FlatBufferEnumAttribute.cs
