+++ src/FlatSharp.Runtime/Attributes/FlatBufferStructAttribute.cs
