+++ src/FlatSharp.Runtime/Attributes/FlatBufferTableAttribute.cs
