+++ src/FlatSharp.Runtime/Attributes/FlatSharpGeneratedSerializerAttribute.cs
