+++ src/FlatSharp.Runtime/BufferTooSmallException.cs
