+++ src/FlatSharp.Runtime/FlatBufferCacheVector.cs
