+++ src/FlatSharp.Runtime/IGeneratedSerializer.cs
