+++ src/FlatSharp.Runtime/IO/ArrayInputBuffer.cs
