+++ src/FlatSharp.Runtime/IO/InputBuffer.cs
