+++ src/FlatSharp.Runtime/IO/MemoryInputBuffer.cs
