+++ src/FlatSharp.Runtime/IO/MemoryTypeChanger.cs
