+++ src/FlatSharp.Runtime/IO/SpanWriter.cs
