+++ src/FlatSharp.Runtime/NotMutableException.cs
