+++ src/FlatSharp.Runtime/SerializationContext.cs
