+++ src/FlatSharp.Runtime/SerializationHelpers.cs
