+++ src/FlatSharp.Runtime/UnionTypes.tt
