+++ src/FlatSharp.Runtime/VTableBuilder.cs
