+++ src/FlatSharp.Unsafe/FlatSharp.Unsafe.csproj
@@ -14,6 +14,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FlatSharp\FlatSharp.csproj" />
+    <ProjectReference Include="..\FlatSharp.Runtime\FlatSharp.Runtime.csproj" />
   </ItemGroup>
 </Project>
