+++ src/FlatSharp.sln
@@ -17,6 +17,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FlatSharp.Unsafe", "FlatSha
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FlatSharp.Compiler", "FlatSharp.Compiler\FlatSharp.Compiler.csproj", "{9B3C0F61-A742-4D5A-A487-7D4DB914EC39}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FlatSharp.Runtime", "FlatSharp.Runtime\FlatSharp.Runtime.csproj", "{07B42C28-D362-49D5-91D6-795634FF49AA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +53,10 @@ Global
 		{9B3C0F61-A742-4D5A-A487-7D4DB914EC39}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9B3C0F61-A742-4D5A-A487-7D4DB914EC39}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9B3C0F61-A742-4D5A-A487-7D4DB914EC39}.Release|Any CPU.Build.0 = Release|Any CPU
+		{07B42C28-D362-49D5-91D6-795634FF49AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07B42C28-D362-49D5-91D6-795634FF49AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07B42C28-D362-49D5-91D6-795634FF49AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07B42C28-D362-49D5-91D6-795634FF49AA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
