+++ src/FlatSharp/Serialization/RoslynSerializerGenerator.cs
@@ -122,6 +122,7 @@ $@"
                 MetadataReference.CreateFromFile(typeof(List<int>).Assembly.Location),
                 MetadataReference.CreateFromFile(typeof(System.Collections.ArrayList).Assembly.Location),
                 MetadataReference.CreateFromFile(typeof(ValueType).Assembly.Location),
+                MetadataReference.CreateFromFile(typeof(IGeneratedSerializer<byte>).Assembly.Location),
                 MetadataReference.CreateFromFile(Assembly.Load("netstandard").Location),
                 MetadataReference.CreateFromFile(typeof(System.IO.InvalidDataException).Assembly.Location),
                 MetadataReference.CreateFromFile(Assembly.Load("System.Runtime").Location),
