+++ src/FlatSharpTests/FlatSharpTests.csproj
@@ -22,6 +22,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\FlatSharp.Compiler\FlatSharp.Compiler.csproj" />
+    <ProjectReference Include="..\FlatSharp.Runtime\FlatSharp.Runtime.csproj" />
     <ProjectReference Include="..\FlatSharp.Unsafe\FlatSharp.Unsafe.csproj" />
     <ProjectReference Include="..\FlatSharp\FlatSharp.csproj" />
     <ProjectReference Include="..\Google.FlatBuffers\Google.FlatBuffers.csproj" />
